refactor model training: preprocessing and data import
