+++ datamodels/wrappers/feature_extension/__init__.py
@@ -1,7 +1,3 @@
 from .store_interface import StoreInterface
-from .feature_expansion import FeatureExpansion
-from .polynomial_expansion import PolynomialExpansion
-from .spline_interpolator import SplineInterpolator
-from .identity_expander import IdentityExpander
-from .expander_set import TransformerSet
+from .transformerset import TransformerSet
 from .expandedmodel import ExpandedModel
