+++ datamodels/wrappers/feature_extension/store_interface.py
@@ -1,5 +1,4 @@
 import os
-import joblib
 import pickle
